requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
Werkzeug==2.3.6
SQLAlchemy==2.0.15
